Add specific quantization utils to get around import issues
