+++ tests/models/test_gptq_marlin.py
@@ -12,7 +12,7 @@ import os
 
 import pytest
 
-from tests.utils import is_quant_method_supported
+from tests.quantization.utils import is_quant_method_supported
 from vllm.model_executor.layers.rotary_embedding import _ROPE_DICT
 
 from .utils import check_logprobs_close
