+++ tests/models/test_gptq_marlin_24.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass
 import pytest
 
 from tests.models.utils import check_logprobs_close
-from tests.utils import is_quant_method_supported
+from tests.quantization.utils import is_quant_method_supported
 
 
 @dataclass
