+++ tests/models/test_marlin.py
@@ -14,7 +14,7 @@ from dataclasses import dataclass
 
 import pytest
 
-from tests.utils import is_quant_method_supported
+from tests.quantization.utils import is_quant_method_supported
 
 from .utils import check_logprobs_close
