+++ tests/quantization/test_fp8.py
@@ -5,7 +5,7 @@ Run `pytest tests/quantization/test_fp8.py --forked`.
 import pytest
 import torch
 
-from tests.utils import is_quant_method_supported
+from tests.quantization.utils import is_quant_method_supported
 from vllm.model_executor.layers.quantization.fp8 import Fp8LinearMethod
